STATISTIK-1044 Mergat in ändring från development-branchen: STATISTIK-1040 Hela personnumret används nu vid gruppering av intyg för sjukfallsberäkningar

+++ statistik-service/src/main/java/se/inera/statistics/service/warehouse/SjukfallUtil.java
@@ -110,12 +110,12 @@ public class SjukfallUtil {
 
     private Collection<Sjukfall> calculateSjukfall(Aisle aisle, Predicate<Fact> filter, int cutoff) {
         Collection<Sjukfall> sjukfalls = new ArrayList<>();
-        Map<Integer, Sjukfall> active = new HashMap<>();
+        Map<Long, Sjukfall> active = new HashMap<>();
         for (Fact line : aisle) {
             if (line.getStartdatum() >= cutoff) {
                 break;
             }
-            int key = line.getPatient();
+            long key = line.getPatient();
             Sjukfall sjukfall = active.get(key);
 
             if (sjukfall == null) {
